Reorganize
Move matrix stuff to its own package.
Rename DuskClient to LoggerBackend and remove the connect method. It may
be useful to have more than one kind of logger behavior, but not more
than one type of Dusk client.
Rename NetworkedDuskClient to DuskClient since it's the only one now.
Remove DuskClient tests in MecanumDrive. We'll find a better place to
test it.
