+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/localization/LocalizationData.java
@@ -2,6 +2,8 @@ package org.firstinspires.ftc.teamcode.localization;
 
 import java.util.List;
 
+import org.firstinspires.ftc.teamcode.matrix.Vec2;
+
 public interface LocalizationData {
     double getPositionProbability(Vec2 pos);
