+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/localization/RobotLocalizer.java
@@ -1,5 +1,8 @@
 package org.firstinspires.ftc.teamcode.localization;
 
+import org.firstinspires.ftc.teamcode.matrix.Mat3;
+import org.firstinspires.ftc.teamcode.matrix.Vec2;
+
 public interface RobotLocalizer {
     void invalidateCache();
