+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/logging/Log.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.dusk;
+package org.firstinspires.ftc.teamcode.logging;
 
 public interface Log {
     byte[] serialize();
