+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/matrix/Mat2.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.localization;
+package org.firstinspires.ftc.teamcode.matrix;
 
 public final class Mat2 {
     private final double[] mat;
