+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/matrix/Mat3.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.localization;
+package org.firstinspires.ftc.teamcode.matrix;
 
 public final class Mat3 {
     private final double[] mat;
